requirements.txt
Flask==2.3.2
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
Pillow==10.0.0
scikit-learn==1.2.2
matplotlib==3.7.1
